Small clean up
* remove one off methods
* make some instance variables a little more private
* remove unnecessary instance variable in ProtegeServerState
